+++ Showcase/Showcase/Pages/Tokens/Dimension/Size/SizeTokenPage.swift
@@ -59,6 +59,9 @@ struct SizeTokenPage: View {
 
     var body: some View {
         Group {
+            Section {
+                ShowcaseTokenCode(code: "theme.sizeIconWithHeadingXLargeShort.dimension(for: horizontalSizeClass ?? .regular)")
+            }
             Section { illustrationForIconDecorative() } header: {
                 header("app_tokens_dimension_size_iconDecorative_label")
             }